requirements.txt
Flask==1.1.4
flask-mongoengine==1.0.0
mongomock==3.23.0
nose==1.3.7
flask-restful==0.3.9
flask-cors==3.0.10
flask-jwt-extended==4.2.3
requests==2.26.0
requests_toolbelt==0.9.1
emoji==1.4.2
bleach<4.0
boto3==1.18.35
Pillow==8.3.1
sentry-sdk[flask]==1.3.1

gunicorn==20.1.0
dnspython==2.1.0
